LayoutTests/media/track/track-webvtt-tc001-utf8-expected.txt

Tests that UTF-8 encoded characters are recognized properly and that different encodings (iconv) are not recognized as a WebVTT file (we do allow it, it just looks ugly).


*** Testing text track 0
EXPECTED (cues.length == '2') OK
EXPECTED (cues[i].id == '1') OK
EXPECTED (cues[i].startTime == '0') OK
EXPECTED (cues[i].endTime == '30.5') OK
EXPECTED (cues[i].getCueAsSource() == '景気判断') OK
EXPECTED (cues[i].id == '2') OK
EXPECTED (cues[i].startTime == '31') OK
EXPECTED (cues[i].endTime == '1200.5') OK
EXPECTED (cues[i].getCueAsSource() == '電力不足') OK

*** Testing text track 1
EXPECTED (cues.length == '2') OK
END OF TEST
